backend/pyproject.toml
[project]
name = "documentcheck"
version = "0.1.0"
description = "围标检测系统 - 投标文件分析与围标串标行为检测"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    # Document parsing
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    # C5 parser-pipeline: 图片指纹 + lxml 文本框 xpath
    "Pillow>=10.0",
    "imagehash>=4.3",
    "lxml>=5.0",
    # C4 file-upload: 压缩包/文件类型/编码探测
    "py7zr>=0.21",
    "rarfile>=4.2",
    "chardet>=5.2",
    # Windows 用 python-magic-bin(自带 libmagic);Linux 用 python-magic + libmagic1
    'python-magic-bin>=0.4 ; sys_platform == "win32"',
    'python-magic>=0.4 ; sys_platform != "win32"',
    # Text analysis
    "jieba>=0.42",
    "scikit-learn>=1.5",
    "pandas>=2.2",
    "numpy>=1.26",
    # Utils
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # passlib 1.7.4 的 bcrypt backend self-test 与 bcrypt>=4.1 不兼容(72 字节 API 变化)
    # 2026-04-14 C2 apply 阶段发现,pin 到 4.0.x 以保持兼容
    "bcrypt>=4.0,<4.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
]
ocr = [
    "paddleocr>=2.7",
    "paddlepaddle>=2.6",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
# session-scoped event loop: 让共享的 async engine 在所有 test 间复用,避免
# "Event loop is closed" 错。fixture 默认也随之提升到 session scope。
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
